+++ src/routes/tasks/editTask/[taskId]/+page.svelte
@@ -1,15 +1,48 @@
 <script lang="ts">
-    import { TaskType, type TaskItem } from "$lib/components/schedule/task/TaskItem";
+    import { page } from '$app/stores';
     import { goto } from '$app/navigation';
-	import { taskStore } from "$components/schedule/task/taskStore";
+    import { taskStore } from '$components/schedule/task/taskStore';
+    import { TaskType } from '$components/schedule/task/TaskItem';
+    import type { TaskItem } from '$components/schedule/task/TaskItem';
 
-    function onCancel() {
+    // Extract taskId from the URL and get the referenced task
+	const taskId = $page.params.taskId;
+    let taskItem = $taskStore.find(x => x.id === Number(taskId));
+    
+    if (taskItem == undefined) {
+        alert("Το συμβάν δεν βρέθηκε.");
         goto('/schedule/tasks');
     }
 
-    function onSubmit(event: Event) {
-        event.preventDefault();
+    let startDate = new Date(taskItem?.date.startDate || new Date());
+    let start =  startDate.getFullYear() + "-" + 
+                (String(startDate.getMonth() + 1)).padStart(2, '0') + "-" + 
+                String(startDate.getDate()).padStart(2, '0') + "T" + 
+                String(startDate.getHours()).padStart(2, '0') + ":" + 
+                String(startDate.getMinutes()).padStart(2, '0');
 
+    let endDate = new Date(taskItem?.date.endDate || new Date());
+    let end =  endDate.getFullYear() + "-" + 
+            (String(endDate.getMonth() + 1)).padStart(2, '0') + "-" + 
+            String(endDate.getDate()).padStart(2, '0') + "T" + 
+            String(endDate.getHours()).padStart(2, '0') + ":" + 
+            String(endDate.getMinutes()).padStart(2, '0');
+
+
+    function onCancel() {
+        goto('/tasks');
+    }
+
+    function onDelete() {
+        taskStore.update(oldArray => {
+            // Filter out the item with the specified id
+            const newArray = oldArray.filter(item => item.id !== taskItem?.id);
+            return newArray;
+        });
+        goto('/tasks');
+    }
+
+    function onSubmit() {
         const startInputElement = document.getElementById("start") as HTMLIonDatetimeElement;
         const endInputElement = document.getElementById("end") as HTMLIonDatetimeElement;
         const startDate = new Date(startInputElement.value?.toString() || new Date().toString());
@@ -21,7 +54,7 @@
         }
 
         let formData: TaskItem = {
-            id: new Date().getTime(),
+            id: taskItem?.id || new Date().getTime(),
             title: (document.getElementById("title") as HTMLInputElement)?.value || "Ανώνυμο συμβάν",
             description: (document.getElementById("description") as HTMLInputElement)?.value || "Χωρίς περιγραφή",
             date: {
@@ -31,14 +64,29 @@
             type: ((document.getElementById("type") as HTMLIonRadioGroupElement)?.value as TaskType) || TaskType.OTHER
         };
 
-        $taskStore = $taskStore.concat(formData);
-        goto('/schedule/tasks');
+        if (formData == taskItem) {
+            goto('/tasks');
+            return;
+        }
+
+        taskStore.update(oldArray => {
+            // Find the index of the item with the specified id
+            const index = oldArray.findIndex(item => item.id === taskItem?.id);
+
+            // If the item is found, update its value
+            if (index !== -1) {
+                oldArray[index] = formData;
+            }
+            return [...oldArray];
+        });
+
+        goto('/tasks');
     }
 </script>
 
 <ion-header>
     <ion-toolbar>
-        <ion-title>Νέο συμβάν</ion-title>
+        <ion-title>Επεξεργασία {taskItem?.type == TaskType.PROJECT ? 'εργασίας' : taskItem?.type == TaskType.TEST ? 'προόδου' : 'συμβάντος'}</ion-title>
     </ion-toolbar>
 </ion-header>
 
@@ -49,6 +97,7 @@
             label="Τίτλος"
             label-placement="stacked"
             id="title"
+            value={taskItem?.title}
             type="text"
             contenteditable="true"
             spellcheck={true}
@@ -59,6 +108,7 @@
             label="Περιγραφή"
             label-placement="stacked"
             id="description"
+            value={taskItem?.description}
             type="text"
             contenteditable="true"
             spellcheck={true}
@@ -68,7 +118,7 @@
             <ion-label style="flex: 1;">Από</ion-label>
             <ion-datetime-button datetime="start"></ion-datetime-button>
             <ion-modal keep-contents-mounted={true}>
-                <ion-datetime id="start" locale="el-GR"></ion-datetime>
+                <ion-datetime id="start" locale="el-GR" value={start}></ion-datetime>
             </ion-modal>
         </ion-div>
         
@@ -76,13 +126,13 @@
             <ion-label style="flex: 1;">Έως</ion-label>
             <ion-datetime-button datetime="end"/>
             <ion-modal keep-contents-mounted={true}>
-                <ion-datetime id="end" locale="el-GR"></ion-datetime>
+                <ion-datetime id="end" locale="el-GR" value={end}></ion-datetime>
             </ion-modal>
         </ion-div>
         
         
         <ion-row class="expanded-row">
-            <ion-radio-group id="type" value="other" class="expanded-radio-group">
+            <ion-radio-group id="type" value={taskItem?.type.valueOf()} class="expanded-radio-group">
                 <ion-radio mode="ios" slot="end" value="test">Πρόοδος</ion-radio>
                 <ion-radio mode="ios" slot="end" value="project">Εργασία</ion-radio>
                 <ion-radio mode="ios" slot="end" value="other">Άλλο</ion-radio>
@@ -90,8 +140,9 @@
         </ion-row>
         
         <div style="display: flex; justify-content: space-between; padding-top: 5%">
-            <ion-button type="reset" on:ionFocus={onCancel} color="light">ΑΚΥΡΟ</ion-button>
-            <ion-button type="submit">ΠΡΟΣΘΗΚΗ</ion-button>  
+            <ion-button type="reset" color="light" on:ionFocus={onCancel}>ΑΚΥΡΟ</ion-button>
+            <ion-button type="button" color="secondary" on:ionFocus={onDelete}>ΔΙΑΓΡΑΦΗ</ion-button>
+            <ion-button type="submit" on:ionFocus={onSubmit}>ΕΝΗΜΕΡΩΣΗ</ion-button>
         </div>
     </form>
 </ion-content>
